Copy in the inverse code
