+++ test.py
@@ -1,36 +1,50 @@
 # -*- coding: utf-8 -*-
+"""
+Created on Tue Jun 27 13:39:13 2017
+
+@author: gbaechle
+"""
+
+import numpy as np
+import scipy as sp
+import scipy.optimize
+
+N = 100
+
+H = np.random.rand(N, N) + 1j * np.random.rand(N, N)
+Hr = np.real(H)
+Hi = np.imag(H)
+
+p = np.random.rand(N)
+
+i = np.abs(H @ p)**2
+
+#print(i)
+#print((Hr @ p)**2 + (Hi @ p)**2)
+
+p_est = np.random.rand(N)
 
 
-import matplotlib.pyplot as plt
-import seaborn as sns
+for n in range(10):
+    
+    p_est =  np.linalg.lstsq( Hr, np.sqrt( i-(Hi @ p_est)**2 ) )[0]
+    print(p_est)
+    p_est =  np.linalg.lstsq( Hi, np.sqrt( i-(Hr @ p_est)**2 ) )[0]
+    print(p_est)
+    
+print('------------')
+print(p)
 
-import lippmann as lip1
-import lippmann2 as lip2
-from tools import *
+diff = lambda x: (Hr @ x)**2 + (Hi @ x)**2 - i
+diff = lambda x: np.abs(H @ x)**2 - i
+p_est = sp.optimize.least_squares(diff, np.random.rand(N)).x
 
-sns.set_palette("Blues_r")
-sns.set_style("whitegrid")
-
-sns.set_context("notebook", font_scale=1.0, rc={"lines.linewidth": 3.0})
+print(p_est)
 
 
-#close all figures
-plt.close("all")
-gui_manager = None
 
-path_CAVE = '/Users/gbaechle/EPFL/PhD/BRDF Data and Rendering Engines/CAVE - multispectral image database/fake_and_real_strawberries_ms/fake_and_real_strawberries_ms'
-
-lippmann_plate = load_multispectral_image_CAVE(path_CAVE)
-spectrum = lippmann_plate.spectrums
-
-lippmann = lip2.LippmannContinuous(spectrum.wave_lengths, 10, 10)
-
-print(lippmann.phases().shape)
-print(lippmann.complex_amplitude().shape)
-
-print(lippmann.compute_wave_function(np.linspace(0, 100E-9, 100)).shape)
-
-print(lippmann.I.shape)
-print(lippmann.replay())
-
-lippmann_discrete = lippmann.to_uniform_freq(34)
+    
+    
+    
+    
+
